pyproject.toml
[tool.poetry]
name = "django-lettings-cicd"
version = "0.1.0"
description = "Orange County Lettings website"
authors = ["Anna <anna@orangecountylettings.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.1"
sentry-sdk = "1.39.1"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
flake8 = "3.7.0"
pytest-django = "4.7.0"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.6.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "oc_lettings_site.settings"
addopts = "-v --cov --cov-report term  --cov-report html --no-cov-on-fail --cov-fail-under 80"
[tool.coverage.run]
source = ["oc_lettings_site", "lettings", "profiles"]
omit = ["tests.py", "*/tests/*", "*/migrations/*", "*/asgi.py", "*/wsgi.py"]
